Changed CardRepositoryImpl, now internal in CardRepository.kt
